Some cleanup done
